refactor: 💡 refactor code
