+++ src/api/member/index.ts
@@ -1,6 +1,5 @@
 import { get } from '../../common/services'
-import { GetMemberInfoRes, GetMemberInfoReq, GetgraduateMemberInfoRes, GetgraduateMemberReq} from './interface'
-import {GetMemberMessage } from './interface'
+import { GetMemberInfoRes, GetMemberInfoReq, GetgraduateMemberInfoRes, GetgraduateMemberReq, GetMemberMessage } from './interface'
 
 /**
  * @param team 组别
@@ -10,4 +9,3 @@ import {GetMemberMessage } from './interface'
 export const getMemberInfo = (params: GetMemberInfoReq) => get<GetMemberInfoRes>(`/api/member/list/current/${params.team}/1?size=${params.size}`)
 export const getgraduateMemberInfo = (params: GetgraduateMemberReq) => get<GetgraduateMemberInfoRes>(`/api/member/list/graduate/${params.year}/1?size=${params.size}`)
 export const getMemberDetail = (username: string) => get<GetMemberMessage>(`/api/member/detail/${username}`)
-
