+++ src/api/wiki/index.ts
@@ -1,43 +1,12 @@
-export interface IWikiItem {
-  id: number
-  author: {
-      id: number
-      realName: string
-      userName: string
-      team: string
-      portrait: string
-  }
-  title: string
-  img: string
-  summary: string
-  type: {
-      id: number
-      typeName: string
-  }
-  pubTime: string
-  explore: number
+import { get } from '../../common/services'
+import { FetchWikiListReq, IWikiItem, IWikeDetail } from './interface'
+import { IPageInfo } from '../../common/Inteface';
+
+export interface IGroupType {
+  count: number,
+  id: number,
+  typeName: string
 }
-export interface FetchWikiListReq {
-  size: number;
-  pageNum?:number;
-  type?:string;
-}
-export interface IWikeDetail{
-        content: string,
-        id: number,
-        author: {
-            id: number,
-            realName: string,
-            userName:string,
-            team:string
-        },
-        title: string,
-        img:string,
-        summary:string,
-        type: {
-            id: number,
-            typeName: string
-        },
-        pubTime: string,
-        explore: number
-}
+export const fetchWikiList = (req: FetchWikiListReq) => get<{ dataList: IWikiItem[] } & { pageInfo: IPageInfo }>(`/api/wiki/list/${(req.type && `type/${req.type}`) || 'all'}/${req.pageNum || 1}?size=${req.size}`)
+export const fetchWikiGroup = () => get<IGroupType[]>('/api/wiki/types');
+export const fetchWikiDetail = (id:string) => get<IWikeDetail>(`/api/wiki/detail/${id}`);
