+++ src/components/header/index.less
@@ -35,7 +35,7 @@
     margin-right: 200px;
   }
   .menu-item {
-    font-size: 15px;
+    font-size: 18px;
     color: white;
     display: inline-block;
   }
