+++ src/pages/home/index.less
@@ -4,15 +4,15 @@
   .carousel-item {
     width: 100%;
     height: 100vh;
-    background-image: url('http://dummyimage.com/200x100');
+    background-image: url('https://mobile.xiyou.edu.cn/src/images/ios.jpg');
     background-size: cover;
     background-repeat: no-repeat;
   }
   .item1{
-    background-image: url('http://dummyimage.com/200x100');
+    background-image: url('https://mobile.xiyou.edu.cn/src/images/ios.jpg');
   }
   .item2{
-    background-image: url('http://dummyimage.com/200x100');
+    background-image: url('https://mobile.xiyou.edu.cn/src/images/ios.jpg');
   }
   .activity,.wiki{
     border-radius: 15px;
