+++ src/pages/home/index.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
 import { Carousel, message } from 'antd'
-import { Link, useHistory } from 'react-router-dom'
 import { fetchActivityList } from '../../api/activity'
 import { ActivityHome } from '../../components/activity-home'
-import { fetchWikiList } from '../../api/wiki/interface'
-import { IWikiItem } from '../../api/wiki'
+import { fetchWikiList } from '../../api/wiki'
+import { IWikiItem } from '../../api/wiki/interface'
 import { GroupIntroduce } from '../../components/group-introduce'
 import { IActivityItem } from '../../api/activity/interface'
 import { WikiHome } from '../../components/wiki-home'
